+++ lib/pragma/operation.rb
@@ -19,6 +19,7 @@ require 'pragma/operation/response/no_content'
 require 'pragma/operation/response/unauthorized'
 require 'pragma/operation/response/service_unavailable'
 require 'pragma/operation/response/internal_server_error'
+require 'pragma/operation/response/payment_required'
 
 module Pragma
   # Operations provide business logic encapsulation for your JSON API.