+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added `Unauthorized` response template
 - Added `ServiceUnavailable` response template
 - Added `InternalServerError` response template
+- Added `PaymentRequired` response template
 
 ### Fixed